src/main/java/com/amnesty/panicbutton/wizard/ViewPagerWithoutSwipe.java
package com.amnesty.panicbutton.wizard;

import android.content.Context;
import android.support.v4.view.ViewPager;
import android.util.AttributeSet;
import android.view.MotionEvent;

public class ViewPagerWithoutSwipe extends ViewPager {
    public ViewPagerWithoutSwipe(Context context) {
        super(context);
    }

    public ViewPagerWithoutSwipe(Context context, AttributeSet attrs) {
        super(context, attrs);
    }

    @Override
    public boolean onTouchEvent(MotionEvent ev) {
        return false;
    }

    @Override
    public boolean onInterceptTouchEvent(MotionEvent ev) {
        return false;
    }
}

